Restructured UTs of bosh and cf
