+++ broker/data-access-layer/bosh/test/bosh.EvaluationContext.spec.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const proxyquire = require('proxyquire');
-const EvaluationContext = proxyquire('../../data-access-layer/bosh/src/EvaluationContext', {
+const EvaluationContext = proxyquire('../src/EvaluationContext', {
   crypto: {
     randomBytes: function (n) {
       return _.fill(new Buffer(n), 120);
