+++ broker/data-access-layer/cf/test/cf.TokenInfo.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const TokenInfo = require('../../data-access-layer/cf/src/TokenInfo');
+const TokenInfo = require('../src/TokenInfo');
 
 const tokenType = 'bearer';
 const expiredToken = 'eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjB9';
